# docs/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Sample API",
    "description" : "server-foundation-template 의 참조 구현.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "tags" : [ {
    "name" : "User",
    "description" : "사용자"
  } ],
  "paths" : {
    "/api/v1/users" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "사용자 목록 조회",
        "operationId" : "list",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PageQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "X-Trace-Id" : {
                "description" : "이 요청의 traceId. meta.traceId, 로그의 traceId 와 같은 값이다.",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "description" : "2xx 면 true. HTTP 상태와 항상 일치한다."
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/PageResponseUserResponse"
                    },
                    "error" : {
                      "$ref" : "#/components/schemas/ApiError"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required" : [ "data", "error", "meta", "success" ]
                }
              }
            }
          },
          "400" : {
            "description" : "COMMON_INVALID_PARAMETER — 요청 값이 올바르지 않습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401" : {
            "description" : "COMMON_UNAUTHORIZED — 인증이 필요합니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "403" : {
            "description" : "COMMON_FORBIDDEN — 접근 권한이 없습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "500" : {
            "description" : "COMMON_INTERNAL_ERROR — 일시적인 오류가 발생했습니다. traceId 를 담당자에게 알려 주세요.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "User" ],
        "summary" : "사용자 등록",
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "headers" : {
              "X-Trace-Id" : {
                "description" : "이 요청의 traceId. meta.traceId, 로그의 traceId 와 같은 값이다.",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "description" : "2xx 면 true. HTTP 상태와 항상 일치한다."
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/UserResponse"
                    },
                    "error" : {
                      "$ref" : "#/components/schemas/ApiError"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required" : [ "data", "error", "meta", "success" ]
                }
              }
            }
          },
          "409" : {
            "description" : "USER_EMAIL_DUPLICATED — 이미 등록된 이메일입니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "COMMON_INVALID_PARAMETER — 요청 값이 올바르지 않습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401" : {
            "description" : "COMMON_UNAUTHORIZED — 인증이 필요합니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "403" : {
            "description" : "COMMON_FORBIDDEN — 접근 권한이 없습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "500" : {
            "description" : "COMMON_INTERNAL_ERROR — 일시적인 오류가 발생했습니다. traceId 를 담당자에게 알려 주세요.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "사용자 단건 조회",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "X-Trace-Id" : {
                "description" : "이 요청의 traceId. meta.traceId, 로그의 traceId 와 같은 값이다.",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "description" : "2xx 면 true. HTTP 상태와 항상 일치한다."
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/UserResponse"
                    },
                    "error" : {
                      "$ref" : "#/components/schemas/ApiError"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required" : [ "data", "error", "meta", "success" ]
                }
              }
            }
          },
          "404" : {
            "description" : "USER_NOT_FOUND — 사용자를 찾을 수 없습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "COMMON_INVALID_PARAMETER — 요청 값이 올바르지 않습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401" : {
            "description" : "COMMON_UNAUTHORIZED — 인증이 필요합니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "403" : {
            "description" : "COMMON_FORBIDDEN — 접근 권한이 없습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "500" : {
            "description" : "COMMON_INTERNAL_ERROR — 일시적인 오류가 발생했습니다. traceId 를 담당자에게 알려 주세요.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "User" ],
        "summary" : "사용자 삭제",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content",
            "headers" : {
              "X-Trace-Id" : {
                "description" : "이 요청의 traceId. meta.traceId, 로그의 traceId 와 같은 값이다.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "USER_NOT_FOUND — 사용자를 찾을 수 없습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "COMMON_INVALID_PARAMETER — 요청 값이 올바르지 않습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401" : {
            "description" : "COMMON_UNAUTHORIZED — 인증이 필요합니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "403" : {
            "description" : "COMMON_FORBIDDEN — 접근 권한이 없습니다.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "500" : {
            "description" : "COMMON_INTERNAL_ERROR — 일시적인 오류가 발생했습니다. traceId 를 담당자에게 알려 주세요.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "CreateUserRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "maxLength" : 2147483647,
            "minLength" : 8
          }
        },
        "required" : [ "email", "password" ]
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "email" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "email", "userId" ]
      },
      "PageQuery" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "page", "size" ]
      },
      "PageMeta" : {
        "type" : "object",
        "properties" : {
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "number", "size", "totalElements", "totalPages" ]
      },
      "PageResponseUserResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMeta"
          }
        },
        "required" : [ "content", "page" ]
      },
      "ApiError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "클라이언트가 분기에 쓰는 값"
          },
          "message" : {
            "type" : "string",
            "description" : "사용자에게 보여줄 문장"
          },
          "details" : {
            "type" : "array",
            "description" : "필드 단위 사유. 검증 실패에서만 채운다.",
            "items" : {
              "$ref" : "#/components/schemas/ApiErrorDetail"
            }
          }
        },
        "required" : [ "code", "message" ]
      },
      "ApiErrorDetail" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "실패한 필드"
          },
          "reason" : {
            "type" : "string",
            "description" : "실패 사유"
          }
        },
        "required" : [ "field", "reason" ]
      },
      "ApiResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "data" : { },
          "error" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/ApiError"
            }, {
              "type" : "null"
            } ]
          },
          "meta" : {
            "$ref" : "#/components/schemas/ResponseMeta"
          }
        },
        "required" : [ "meta", "success" ]
      },
      "ResponseMeta" : {
        "type" : "object",
        "properties" : {
          "traceId" : {
            "type" : "string",
            "description" : "응답 헤더 X-Trace-Id, 로그의 traceId 와 같은 값"
          },
          "timestamp" : {
            "type" : "string",
            "description" : "응답 생성 시각 (UTC, 밀리초)"
          },
          "path" : {
            "type" : "string",
            "description" : "요청 경로"
          }
        },
        "required" : [ "path", "timestamp", "traceId" ]
      },
      "ApiResponseVoid" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean",
            "description" : "2xx 면 true. HTTP 상태와 항상 일치한다."
          },
          "data" : {
            "description" : "실패 응답의 data 는 항상 null"
          },
          "error" : {
            "$ref" : "#/components/schemas/ApiError"
          },
          "meta" : {
            "$ref" : "#/components/schemas/ResponseMeta"
          }
        },
        "required" : [ "data", "error", "meta", "success" ]
      }
    }
  }
}